Rename content status mapping module, parse correct dataset item
